reporting/document-reporting-bypass-report-to.https.sub.html

<!DOCTYPE HTML>
<html>

<head>
  <title>Test that reports ignore Report-To header when Reporting-Endpoints is configured</title>
  <script src='/resources/testharness.js'></script>
  <script src='/resources/testharnessreport.js'></script>
  <script src='resources/report-helper.js'></script>
</head>

<body>
  <script>
    promise_test(async t => {
      return new Promise(resolve => {
        new ReportingObserver((reports, observer) => resolve(reports),
          { types: ['document-policy-violation'] }).observe();
      }).then((reports) => {
        assert_equals(reports[0].type, 'document-policy-violation');
      })
    }, "document policy violation observed");
  </script>
  <script>document.write("This should be written into the document");</script>
  <script>
    const base_url = `${location.protocol}//${location.host}`;
    const endpoint = `${base_url}/reporting/resources/report.py`;
    const report_to_id = 'd0d517bf-891b-457a-b970-8b2b2c81a0bf';
    const reporting_endpoints_id = '1e1f1d95-a52d-489c-b688-2b52b80f1a81';
    promise_test(async t => {
      await wait(3000);
      // Verify no reports sent to Report-To endpoint
      let reports = await pollReports(endpoint, report_to_id);
      assert_equals(reports.length, 0);
      // Verify report is received on Reporting-Endpoints endpoint
      reports = await pollReports(endpoint, reporting_endpoints_id);
      checkReportExists(reports, 'document-policy-violation', location.href);
    }, "Only the Reporting-Endpoints configured endpoint received reports.");
  </script>

</body>

</html>
